Split the `Item` model to `Update` & `Add` different models.
