+++ src/Todoist.Net/Models/Item.cs
@@ -1,6 +1,4 @@
 using System;
-using System.Collections.Generic;
-using System.Collections.ObjectModel;
 using System.Text.Json.Serialization;
 
 namespace Todoist.Net.Models
@@ -8,32 +6,11 @@ namespace Todoist.Net.Models
     /// <summary>
     /// Represents a Todoist task.
     /// </summary>
-    public class Item : BaseEntity, IWithRelationsArgument
+    public class Item : UpdateItem
     {
-        /// <summary>
-        /// Initializes a new instance of the <see cref="Item" /> class.
-        /// </summary>
-        /// <param name="content">The content.</param>
-        public Item(string content)
-
-            // ReSharper disable once IntroduceOptionalParameters.Global
-            : this(content, default)
-        {
-        }
-
-        /// <summary>
-        /// Initializes a new instance of the <see cref="Item" /> class.
-        /// </summary>
-        /// <param name="content">The content.</param>
-        /// <param name="projectId">The project identifier.</param>
-        public Item(string content, ComplexId projectId)
-        {
-            Content = content;
-            ProjectId = projectId;
-        }
-
         [JsonConstructor]
         internal Item()
+            : base(default)
         {
         }
 
@@ -45,13 +22,6 @@ namespace Todoist.Net.Models
         public string AddedByUid { get; set; }
 
         /// <summary>
-        /// Gets or sets the assigned by uid.
-        /// </summary>
-        /// <value>The assigned by uid.</value>
-        [JsonPropertyName("assigned_by_uid")]
-        public string AssignedByUid { get; set; }
-
-        /// <summary>
         /// Gets or sets order of project. Defines the position of the project among all the projects with the same parent_id.
         /// </summary>
         /// <value>The project order.</value>
@@ -59,61 +29,11 @@ namespace Todoist.Net.Models
         public int? ChildOrder { get; set; }
 
         /// <summary>
-        /// Gets or sets a value indicating whether this <see cref="Item" /> is collapsed.
-        /// </summary>
-        /// <value><c>null</c> if [collapsed] contains no value, <c>true</c> if [collapsed]; otherwise, <c>false</c>.</value>
-        [JsonPropertyName("collapsed")]
-        public bool? Collapsed { get; set; }
-
-        /// <summary>
-        /// Gets or sets the content.
-        /// </summary>
-        /// <value>The content.</value>
-        [JsonPropertyName("content")]
-        public string Content { get; set; }
-
-        /// <summary>
         /// Gets the date added.
         /// </summary>
         /// <value>The date added.</value>
         [JsonPropertyName("added_at")]
         public DateTime? AddedAt { get; internal set; }
-
-        /// <summary>
-        /// Gets or sets the day order.
-        /// </summary>
-        /// <value>The day order.</value>
-        [JsonPropertyName("day_order")]
-        public int? DayOrder { get; set; }
-
-        /// <summary>
-        /// Gets or sets the description.
-        /// </summary>
-        /// <value>The description.</value>
-        [JsonPropertyName("description")]
-        public string Description { get; set; }
-
-        /// <summary>
-        /// Gets or sets the due date.
-        /// </summary>
-        /// <value>
-        /// The due date.
-        /// </value>
-        [JsonPropertyName("due")]
-        public DueDate DueDate { get; set; }
-
-        /// <summary>
-        /// Gets or sets the duration.
-        /// </summary>
-        /// <remarks>
-        /// Durations are only available for Todoist Premium users.
-        /// </remarks>
-        /// <value>
-        /// The duration.
-        /// </value>
-        [JsonPropertyName("duration")]
-        [JsonIgnore(Condition = JsonIgnoreCondition.Never)]
-        public Duration Duration { get; set; }
 
         /// <summary>
         /// Gets a value indicating whether this instance is checked.
@@ -130,13 +50,6 @@ namespace Todoist.Net.Models
         public bool? IsDeleted { get; internal set; }
 
         /// <summary>
-        /// Gets the labels.
-        /// </summary>
-        /// <value>The labels.</value>
-        [JsonPropertyName("labels")]
-        public ICollection<string> Labels { get; internal set; }
-
-        /// <summary>
         /// Gets or sets the id of the parent task. Set to <see langword="null" /> for root tasks.
         /// </summary>
         /// <value>
@@ -146,25 +59,11 @@ namespace Todoist.Net.Models
         public string ParentId { get; set; }
 
         /// <summary>
-        /// Gets or sets the priority.
-        /// </summary>
-        /// <value>The priority.</value>
-        [JsonPropertyName("priority")]
-        public Priority? Priority { get; set; }
-
-        /// <summary>
         /// Gets or sets the project identifier.
         /// </summary>
         /// <value>The project identifier.</value>
         [JsonPropertyName("project_id")]
         public ComplexId? ProjectId { get; set; }
-
-        /// <summary>
-        /// Gets or sets the responsible uid.
-        /// </summary>
-        /// <value>The responsible uid.</value>
-        [JsonPropertyName("responsible_uid")]
-        public string ResponsibleUid { get; set; }
 
         /// <summary>
         /// Gets or sets section of project. Defines the section that the task belongs to.
@@ -188,16 +87,5 @@ namespace Todoist.Net.Models
         [JsonPropertyName("user_id")]
         public string UserId { get; internal set; }
 
-        /// <summary>
-        /// Updates the related temporary ids.
-        /// </summary>
-        /// <param name="map">The map.</param>
-        void IWithRelationsArgument.UpdateRelatedTempIds(IDictionary<Guid, string> map)
-        {
-            if (ProjectId.HasValue && map.TryGetValue(ProjectId.Value.TempId, out var persistentProjectId))
-            {
-                ProjectId = new ComplexId(persistentProjectId);
-            }
-        }
     }
 }
